+++ src/Symfony/Component/Console/DependencyInjection/AddConsoleCommandPass.php
@@ -12,9 +12,12 @@
 namespace Symfony\Component\Console\DependencyInjection;
 
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\CommandLoader\ContainerCommandLoader;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\Compiler\ServiceLocatorTagPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
+use Symfony\Component\DependencyInjection\TypedReference;
 
 /**
  * Registers console commands.
@@ -23,9 +26,20 @@ use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
  */
 class AddConsoleCommandPass implements CompilerPassInterface
 {
+    private $commandLoaderServiceId;
+    private $commandTag;
+
+    public function __construct($commandLoaderServiceId = 'console.command_loader', $commandTag = 'console.command')
+    {
+        $this->commandLoaderServiceId = $commandLoaderServiceId;
+        $this->commandTag = $commandTag;
+    }
+
     public function process(ContainerBuilder $container)
     {
-        $commandServices = $container->findTaggedServiceIds('console.command', true);
+        $commandServices = $container->findTaggedServiceIds($this->commandTag, true);
+        $lazyCommandMap = array();
+        $lazyCommandRefs = array();
         $serviceIds = array();
 
         foreach ($commandServices as $id => $tags) {
@@ -36,20 +50,51 @@ class AddConsoleCommandPass implements CompilerPassInterface
                 throw new InvalidArgumentException(sprintf('Class "%s" used for service "%s" cannot be found.', $class, $id));
             }
             if (!$r->isSubclassOf(Command::class)) {
-                throw new InvalidArgumentException(sprintf('The service "%s" tagged "console.command" must be a subclass of "%s".', $id, Command::class));
+                throw new InvalidArgumentException(sprintf('The service "%s" tagged "%s" must be a subclass of "%s".', $id, $this->commandTag, Command::class));
             }
 
             $commandId = 'console.command.'.strtolower(str_replace('\\', '_', $class));
-            if ($container->hasAlias($commandId) || isset($serviceIds[$commandId])) {
-                $commandId = $commandId.'_'.$id;
-            }
-            if (!$definition->isPublic()) {
-                $container->setAlias($commandId, $id);
-                $id = $commandId;
+
+            if (!isset($tags[0]['command'])) {
+                if (isset($serviceIds[$commandId]) || $container->hasAlias($commandId)) {
+                    $commandId = $commandId.'_'.$id;
+                }
+                if (!$definition->isPublic()) {
+                    $container->setAlias($commandId, $id);
+                    $id = $commandId;
+                }
+                $serviceIds[$commandId] = $id;
+
+                continue;
             }
 
-            $serviceIds[$commandId] = $id;
+            $serviceIds[$commandId] = false;
+            $commandName = $tags[0]['command'];
+            $lazyCommandMap[$commandName] = $id;
+            $lazyCommandRefs[$id] = new TypedReference($id, $class);
+            $aliases = array();
+
+            foreach ($tags as $tag) {
+                if (!isset($tag['command'])) {
+                    throw new InvalidArgumentException(sprintf('Missing "command" attribute on tag "%s" for service "%s".', $this->commandTag, $id));
+                }
+                if ($commandName !== $tag['command']) {
+                    throw new InvalidArgumentException(sprintf('The "command" attribute must be the same on each "%s" tag for service "%s".', $this->commandTag, $id));
+                }
+                if (isset($tag['alias'])) {
+                    $aliases[] = $tag['alias'];
+                    $lazyCommandMap[$tag['alias']] = $id;
+                }
+            }
+
+            if ($aliases) {
+                $definition->addMethodCall('setAliases', array($aliases));
+            }
         }
+
+        $container
+            ->register($this->commandLoaderServiceId, ContainerCommandLoader::class)
+            ->setArguments(array(ServiceLocatorTagPass::register($container, $lazyCommandRefs), $lazyCommandMap));
 
         $container->setParameter('console.command.ids', $serviceIds);
     }