+++ src/Symfony/Component/Console/Tests/DependencyInjection/AddConsoleCommandPassTest.php
@@ -12,10 +12,13 @@
 namespace Symfony\Component\Console\Tests\DependencyInjection;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Console\CommandLoader\ContainerCommandLoader;
 use Symfony\Component\Console\DependencyInjection\AddConsoleCommandPass;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\DependencyInjection\Argument\ServiceClosureArgument;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
+use Symfony\Component\DependencyInjection\TypedReference;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class AddConsoleCommandPassTest extends TestCase
@@ -53,6 +56,26 @@ class AddConsoleCommandPassTest extends TestCase
         $this->assertSame(array($alias => $id), $container->getParameter('console.command.ids'));
     }
 
+    public function testProcessRegisterLazyCommands()
+    {
+        $container = new ContainerBuilder();
+        $container
+            ->register('my-command', MyCommand::class)
+            ->setPublic(false)
+            ->addTag('console.command', array('command' => 'my:command', 'alias' => 'my:alias'))
+        ;
+
+        (new AddConsoleCommandPass())->process($container);
+
+        $commandLoader = $container->getDefinition('console.command_loader');
+        $commandLocator = $container->getDefinition((string) $commandLoader->getArgument(0));
+
+        $this->assertSame(ContainerCommandLoader::class, $commandLoader->getClass());
+        $this->assertSame(array('my:command' => 'my-command', 'my:alias' => 'my-command'), $commandLoader->getArgument(1));
+        $this->assertEquals(array(array('my-command' => new ServiceClosureArgument(new TypedReference('my-command', MyCommand::class)))), $commandLocator->getArguments());
+        $this->assertSame(array('console.command.symfony_component_console_tests_dependencyinjection_mycommand' => false), $container->getParameter('console.command.ids'));
+    }
+
     public function visibilityProvider()
     {
         return array(
@@ -72,7 +95,7 @@ class AddConsoleCommandPassTest extends TestCase
         $container->addCompilerPass(new AddConsoleCommandPass());
 
         $definition = new Definition('Symfony\Component\Console\Tests\DependencyInjection\MyCommand');
-        $definition->addTag('console.command');
+        $definition->addTag('console.command', array('command' => 'my:command'));
         $definition->setAbstract(true);
         $container->setDefinition('my-command', $definition);
 
@@ -90,7 +113,7 @@ class AddConsoleCommandPassTest extends TestCase
         $container->addCompilerPass(new AddConsoleCommandPass());
 
         $definition = new Definition('SplObjectStorage');
-        $definition->addTag('console.command');
+        $definition->addTag('console.command', array('command' => 'my:command'));
         $container->setDefinition('my-command', $definition);
 
         $container->compile();