+++ src/Symfony/Component/HttpKernel/DataCollector/DataCollector.php
@@ -11,9 +11,10 @@
 
 namespace Symfony\Component\HttpKernel\DataCollector;
 
-use Symfony\Component\VarDumper\Caster\ClassStub;
+use Symfony\Component\VarDumper\Caster\CutStub;
 use Symfony\Component\VarDumper\Cloner\ClonerInterface;
 use Symfony\Component\VarDumper\Cloner\Data;
+use Symfony\Component\VarDumper\Cloner\Stub;
 use Symfony\Component\VarDumper\Cloner\VarCloner;
 
 /**
@@ -31,7 +32,7 @@ abstract class DataCollector implements DataCollectorInterface, \Serializable
     /**
      * @var ClonerInterface
      */
-    private static $cloner;
+    private $cloner;
 
     public function serialize()
     {
@@ -55,15 +56,38 @@ abstract class DataCollector implements DataCollectorInterface, \Serializable
      */
     protected function cloneVar($var)
     {
-        if (null === self::$cloner) {
-            if (!class_exists(ClassStub::class)) {
+        if ($var instanceof Data) {
+            return $var;
+        }
+        if (null === $this->cloner) {
+            if (!class_exists(CutStub::class)) {
                 throw new \LogicException(sprintf('The VarDumper component is needed for the %s() method. Install symfony/var-dumper version 3.4 or above.', __METHOD__));
             }
-
-            self::$cloner = new VarCloner();
-            self::$cloner->setMaxItems(-1);
+            $this->cloner = new VarCloner();
+            $this->cloner->setMaxItems(-1);
+            $this->cloner->addCasters($this->getCasters());
         }
 
-        return self::$cloner->cloneVar($var);
+        return $this->cloner->cloneVar($var);
+    }
+
+    /**
+     * @return callable[] The casters to add to the cloner
+     */
+    protected function getCasters()
+    {
+        return array(
+            '*' => function ($v, array $a, Stub $s, $isNested) {
+                if (!$v instanceof Stub) {
+                    foreach ($a as $k => $v) {
+                        if (is_object($v) && !$v instanceof \DateTimeInterface && !$v instanceof Stub) {
+                            $a[$k] = new CutStub($v);
+                        }
+                    }
+                }
+
+                return $a;
+            },
+        );
     }
 }