+++ src/Symfony/Component/Validator/Tests/Constraints/EqualToValidatorTest.php
@@ -54,6 +54,16 @@ class EqualToValidatorTest extends AbstractComparisonValidatorTestCase
     /**
      * {@inheritdoc}
      */
+    public function provideValidComparisonsToPropertyPath()
+    {
+        return array(
+            array(5),
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function provideInvalidComparisons()
     {
         return array(