+++ src/Symfony/Bridge/Doctrine/CHANGELOG.md
@@ -11,6 +11,11 @@ CHANGELOG
    method throws an exception
  * removed the `DoctrineParserCache` class
 
+3.4.0
+-----
+
+ * added support for doctrine/dbal v2.6 types
+
 3.1.0
 -----