+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -11,6 +11,11 @@ CHANGELOG
 * removed `ConsoleExceptionEvent`
 * removed `ConsoleEvents::EXCEPTION`
 
+3.4.0
+-----
+
+ * added `CommandLoaderInterface` and PSR-11 `ContainerCommandLoader`
+
 3.3.0
 -----